% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_passflow.R
\name{plot_passflow}
\alias{plot_passflow}
\title{Passflow plot function}
\usage{
plot_passflow(pass_data, bin_size = 0)
}
\arguments{
\item{pass_data}{Dataframe that must house pass data only and must have atleast 4 columns: starting x,y and ending x,y.}

\item{bin_size}{Details the binsize the passflow needs to bin to. Default is 20.}
}
\value{
returns a ggplot2 object
}
\description{
This function takes in a dataframe and binsizes
to make a passflow map.
}
\examples{
plot = plot_passflow(data, bin_size=30)
}
