% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shot.R
\name{plot_shot}
\alias{plot_shot}
\title{Function for plotting shots}
\usage{
plot_shot(shotdata, type = "", bin_size = 30, theme = "")
}
\arguments{
\item{shotdata}{Dataframe that houses shot data. Dataframe must contain atleast the following columns: X,Y,xG,result,name}

\item{type}{Type of showcasing the shotmap: hexbin, density, point (default)}

\item{theme}{Theme preferences for display: dark (default), white, rose, almond}
}
\value{
a ggplot2 object
}
\description{
This function allows for data, that has to be scraped from Understat, to be used
for plotting shots.
}
